V0.8 'Implemented RecipeIngredient Relationship'

+++ Application/Recipes/Delete.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -8,12 +9,12 @@ namespace Application.Recipes
 {
     public class Delete
     {
-        public class Command : IRequest<Unit>
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
             public Handler(DataContext context)
@@ -21,15 +22,17 @@ namespace Application.Recipes
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var recipe = await _context.Recipes.FindAsync(request.Id);
 
+                if (recipe == null) return null;
+
                 _context.Recipes.Remove(recipe);
 
-                await _context.SaveChangesAsync();
+                var result = await _context.SaveChangesAsync() > 0;
 
-                return Unit.Value;
+                return result ? Result<Unit>.Success(Unit.Value) : Result<Unit>.Failure("Failed to delete recipe");
             }
         }
     }
